refactored component with tags
