+++ src/layouts/Tile/Tags/MovieDetailsTags/index.js
@@ -1,0 +1,18 @@
+import { Tag, TagContent } from "../styled";
+
+const MovieDetailsTags = ({ genresDetails }) => {
+  return (
+    <>
+      {genresDetails &&
+        genresDetails.map(({ id, name }) => {
+          return (
+            <Tag key={id}>
+              <TagContent movieDetails>{name}</TagContent>
+            </Tag>
+          );
+        })}
+    </>
+  );
+};
+
+export default MovieDetailsTags;
